libs/requirements.txt
acme==0.31.0
aiocontextvars==0.2.2
ansible==2.7.9
asn1crypto==0.24.0
attrs==19.3.0
aws==0.2.5
awsebcli==3.14.11
awspricing==2.0.3
bcrypt==3.1.6
blessed==1.15.0
boto==2.49.0
boto3==1.10.12
botocore==1.13.12
cached-property==1.5.1
cement==2.8.2
certbot==0.31.0
certbot-dns-dnsmadeeasy==0.31.0
certbot-nginx==0.31.0
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
ConfigArgParse==0.14.0
configobj==5.0.6
contextvars==2.4
cryptography==2.5
daiquiri==1.5.0
dataclasses==0.7
defusedxml==0.6.0
Deprecated==1.2.5
distro==1.5.0
dns-lexicon==3.1.5
docker==4.1.0
docopt==0.6.2
docutils==0.14
fabric==2.6.0
fabric2==2.5.0
fbmq==2.5.0
filetype==1.0.5
Flask==1.0.2
future==0.17.1
git-pull-request==2.10.0
gitdb2==2.0.5
GitPython==2.1.11
har2case==0.3.1
httprunner==3.0.0
idna==2.9
imgkit==1.0.2
immutables==0.12
importlib-metadata==1.5.0
invoke==1.3.0
itchat==1.2.32
itsdangerous==1.1.0
Jinja2==2.11.1
jira==2.0.0
jmespath==0.9.3
josepy==1.1.0
jsonpath==0.82
jsonschema==3.2.0
ldap3==2.6.1
loguru==0.4.1
MarkupSafe==1.1.1
mock==2.0.0
multi-key-dict==2.0.3
numpy==1.16.2
oauthlib==3.1.0
pandas==0.24.2
paramiko==2.4.2
parsedatetime==2.4
pathlib==1.0.1
pathlib2==2.3.5
pathspec==0.5.5
pbr==5.1.2
Pillow==5.4.1
pipreqs==0.4.10
poeditor==1.1.2
prettytable==2.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.7
pycparser==2.19
pydantic==1.4
PyGithub==1.43.6
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.15.7
python-dateutil==2.8.0
python-dotenv==0.10.1
python-jenkins==1.4.0
python-ldap==3.2.0
pytz==2018.9
PyYAML==5.3.1
requests==2.23.0
requests-file==1.4.3
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==3.4.2
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
semantic-version==2.5.0
sentry-sdk==0.13.5
sh==1.12.14
six==1.14.0
smmap2==2.0.5
stevedore==1.30.1
termcolor==1.1.0
texttable==0.9.1
tldextract==2.2.0
tornado==6.0.4
urllib3==1.25.8
uWSGI==2.0.18
virtualenv==16.4.3
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.4
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.14.1
wget==3.2
wrapt==1.11.1
wxpy==0.3.9.8
yarg==0.1.9
zipp==3.1.0
zope.component==4.5
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.proxy==4.3.1
